move useHideParameter and ParameterVisibilityToggle close to home
