+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk_setup/components/ParameterSettings/ParameterVisibilityToggle.tsx
@@ -2,7 +2,7 @@ import { t } from "ttag";
 
 import { ActionIcon, Icon, Tooltip } from "metabase/ui";
 
-import type { SdkIframeEmbedSetupExperience } from "../types";
+import type { SdkIframeEmbedSetupExperience } from "../../types";
 
 interface ParameterVisibilityToggleProps {
   parameterName: string;
