+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk_setup/components/ParameterSettings/hooks/use-hide-parameter.ts
@@ -1,4 +1,4 @@
-import { useSdkIframeEmbedSetupContext } from "../context";
+import { useSdkIframeEmbedSetupContext } from "../../../context";
 
 export function useHideParameter() {
   const { settings, updateSettings } = useSdkIframeEmbedSetupContext();
