+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk_setup/components/ParameterSettings/index.tsx
@@ -1,0 +1,1 @@
+export { ParameterSettings } from "./ParameterSettings";
