+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk_setup/hooks/index.ts
@@ -1,4 +1,3 @@
 export { useSdkIframeEmbedNavigation } from "./use-sdk-iframe-embed-navigation";
 export { useRecentItems } from "./use-recent-items";
 export { useParameterList } from "./use-parameter-list";
-export { useHideParameter } from "./use-hide-parameter";
